+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.10.2 (11/24/2018 - )
 * search: update es to 6.5.1
 * kafka: update kafka to 2.1.0
+* redis: multGet values return in same order of keys
 
 ### 6.10.1 (11/16/2018 - 11/23/2018)
 * kafka: set consumer client.id for monitoring